Enhanced _simulate_linear_sem() to support list of SEM types
This commit updates the _simulate_linear_sem function in simulator.py to add more flexibility in defining SEM types.

Previously, the function only accepted a single string SEM type which was assigned to all n noise variables. With this update, it now supports a list of SEM types as well, allowing users to individually specify the SEM type for each noise variable.

If a single string is provided, the function will create a list of size n repeating the same SEM type. If a list is provided, it checks to ensure its size is equal to n and that all elements in the list are strings.

This change enables more nuanced simulations by allowing different SEM types for each variable and enhances the error-checking mechanism to ensure the input is properly formatted.